refactor(libs): refactor auth libs
